vedana-apps/checkin_v4/backend/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic-settings>=2.3.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
alembic>=1.13.1
httpx>=0.27.0
ruff>=0.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
mangum>=0.19.0
email-validator>=2.1.0
python-multipart>=0.0.6
openai>=1.30.0
Pillow>=10.0.0
openpyxl>=3.1.0

